Update structure (#2)
* Update UML

* Rename class

* Renameclass

* Rename class

* Rename class

* Renameclass

* Update structure
