+++ ccms/src/main/java/com/codecool/controller/LogInController.java
@@ -1,0 +1,4 @@
+package com.codecool.controller;
+
+public class LogInController {
+}
