+++ ccms/src/main/java/com/codecool/controller/ManagerController.java
@@ -1,4 +1,4 @@
 package com.codecool.controller;
 
-public class CCmsController {
+public class ManagerController {
 }
