+++ ccms/src/main/java/com/codecool/controller/MentorController.java
@@ -1,0 +1,4 @@
+package com.codecool.controller;
+
+public class MentorController {
+}
