+++ ccms/src/main/java/com/codecool/controller/RegularEmployeeController.java
@@ -1,0 +1,4 @@
+package com.codecool.controller;
+
+public class RegularEmployeeController {
+}
