+++ ccms/src/main/java/com/codecool/model/Assignment.java
@@ -1,4 +1,4 @@
 package com.codecool.model;
 
-public class Assigment {
+public class Assignment {
 }
