+++ ccms/src/main/java/com/codecool/view/StudentMenu.java
@@ -1,17 +1,22 @@
 package com.codecool.view;
 
-public class RegularEmployeeInterface extends Menu{
-    public RegularEmployeeInterface() {
-        this.label = "Employee";
+public class StudentMenu extends Menu {
+    public StudentMenu() {
+        this.label = "Student";
         this.options = new String[] {
-                "List students",
+                "View my grades",
+                "Submit assignment",
         };
     }
 
     protected void executeOption(int option) {
         switch (option) {
             case 1: {
-                System.out.println("List students");
+                System.out.println("View my grades");
+                break;
+            }
+            case 2: {
+                System.out.println("Submit assignment");
                 break;
             }
             default: {
